+++ test/asphalt/core_test.clj
@@ -188,6 +188,14 @@ VALUES (^string $name, ^int $salary, ^string $dept, ^date $joined)")
       u/ds t-insert row)
     (is (= 52 (a/query a/fetch-single-value
                 u/ds t-count [])) "Verify that row was inserted")
+    ;; test fetch-maps
+    (testing "fetch-maps"
+      (let [rows (a/query a/fetch-maps
+                   u/ds t-selfew ["Joe Coder"])]
+        (is (= (zipmap [:name :salary :dept :j_date] vs1) (first rows))))
+      (let [rows (a/query (partial a/fetch-maps {:key-maker a/_label->key})
+                   u/ds t-selfew ["Joe Coder"])]
+        (is (= (zipmap [:name :salary :dept :j-date] vs1) (first rows)))))
     ;; test letcol
     (testing "letcol"
       (let [run-query (fn [row-maker]